% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_lf.R
\name{download_lf}
\alias{download_lf}
\title{Download LANDFIRE Data}
\usage{
download_lf(aoi, time, out_dir)
}
\arguments{
\item{aoi}{SpatVector. The area of interest (AOI) within which you aim to evaluate ERI. Can be defined using \code{terra::vect()}.}

\item{time}{numeric. The total simulated evacuation time, in minutes, within which you aim to evaluate ERI.}

\item{out_dir}{character. The output directory to which LANDFIRE data will be downloaded.}
}
\value{
A list with three SpatRasters: \code{slope} (terrain slope in degrees); \code{elev} (terrain elevation in meters); and \code{veg_type} (LANDFIRE Existing Vegetation Type)
}
\description{
ERI relies on three raster input datasets, each of which are provided by LANDFIRE: (1) slope in degrees; (2) elevation in meters; and (3) existing vegetation type. This function queries LANDFIRE's API with the user-provided area of interest (AOI) and tries to download the three datasets.
}
\details{
\itemize{
\item The ERI algorithm is very processing-intensive. Accordingly, users should avoid using very large AOIs (e.g., >>100 sqkm).
\item In addition to AOI area, the \code{time} argument also partially dictates study area size and processing time. It is used in conjunction with the \code{\link[=slope_rate]{slope_rate()}} function to generate a buffer distance around the AOI within which data will be downloaded. This is to ensure that there are no edge effects in the calculation of ERI within a particular evacuation time frame.
\item Users should specify the same AOI and travel time for both this function and \code{\link[=download_trans]{download_trans()}}.
\item LANDFIRE's API has a small delay betweeen when requests are submitted and when data are returned. This function tries several times to download the returned data, with a 5-second delay between each try. If it reaches 10 tries, it stops, under the assumption that there is a server-side error or delay that is preventing successful data downloading.
\item In the current version of the \code{eri} package, LF2020 data are provided. Note that vegetation may have been disturbed since the release of these data and thus ERI calculations may not reflect updated landscape conditions.
\item Note that this function will save three TIF files to \code{out_dir}: (1) "slope.tif"; (2) "elevation.tif"; and (3) "veg_type.tif".
\item By default, LANDFIRE's API returns data layers in a local Albers projection centered on the centroid of the AOI that maintains both pixel area and N-S-E-W orientation.
}
}
\examples{
# read in aoi
aoi <- vect("C:/temp/study_area.shp")

# download landfire data using a 15 minute evacuation time
lf <- download_lf(aoi, 15, "C:/temp")

# access outputs
slope <- lf$slope
elev <- lf$elev
veg_type <- lf$veg_type
}
