% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trans.R
\name{download_trans}
\alias{download_trans}
\title{Download Transportation Data}
\usage{
download_trans(aoi, time, out_dir)
}
\arguments{
\item{aoi}{SpatVector. The area of interest (AOI) within which you aim to evaluate ERI. Can be defined using \code{terra::vect()}.}

\item{time}{numeric. The total simulated evacuation time, in minutes, within which you aim to evaluate ERI.}

\item{out_dir}{character. The output directory to which transportation data will be downloaded.}
}
\value{
A SpatVector of all transportation features clipped to the extent of the travel time-buffered AOI.
}
\description{
Transportation features (roads, trails, etc.) represent easily walkable portions of the landscape, some of which may be missed in the LANDFIRE Existing Vegetation Type classification. This function downloads transportation linear features (roads, trails, etc.) from OpenStreetMap for use in the ERI modeling process.
}
\details{
\itemize{
\item The ERI algorithm is very processing-intensive. Accordingly, users should avoid using very large AOIs (e.g., >>100 sqkm).
\item In addition to AOI area, the \code{time} argument also partially dictates study area size and processing time. It is used in conjunction with the \code{\link[=slope_rate]{slope_rate()}} function to generate a buffer distance around the AOI within which data will be downloaded. This is to ensure that there are no edge effects in the calculation of ERI within a particular evacuation time frame.
\item Users should specify the same AOI and travel time for both this function and \code{\link[=download_lf]{download_lf()}}.
\item Note that this function will save the following shapefile to \code{out_dir}: "transportation.shp".
\item Although LANDFIRE Existing Vegetation Type captures the dominant land cover conditions, smaller roads, trails, and other minor transportation features may still be classified as restrictive vegetation types (i.e., a trail through the woods being classified as forest). Given the ease with which pedestrians can traverse these constructed features, this function provides the opportunity to download OpenStreetMap transportation linear features. In a subsequent function (\code{\link[=reclass_veg]{reclass_veg()}}), the LANDFIRE Existing Vegetation Type and the OpenStreetMap transportation features are merged to create a combined land cover travel impedance raster layer.
}
}
\examples{
# read in aoi
aoi <- vect("C:/temp/study_area.shp")

# download transportation data using a 15 minute evacuation time
trans <- download_trans(aoi, 15, "C:/temp")
}
