% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eri_area.R
\name{eri_area}
\alias{eri_area}
\title{Calculate ERI for an Area}
\usage{
eri_area(
  veg_type,
  tm,
  aoi,
  time,
  ncores = floor(parallel::detectCores()/2),
  out_dir
)
}
\arguments{
\item{veg_type}{SpatRaster. The OpenStreetMap transportation data-augmented LANDFIRE vegetation type dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}, \code{\link[=download_trans]{download_trans()}}, and \code{\link[=reclass_veg]{reclass_veg()}}.}

\item{tm}{TransitionLayer. The transition matrix generated using \code{\link[=gen_tm]{gen_tm()}}.}

\item{aoi}{SpatVector. The area of interest within which you aim to map ERI.}

\item{time}{numeric. The total simulated evacuation time, in minutes, within which you aim to evaluate ERI.}

\item{ncores}{numeric. The number of cores you want to use to run the algorithm. Defaults to half of the available cores.}

\item{out_dir}{character. The output directory to which the resulting ERI raster will be saved.}
}
\value{
A SpatRaster with four layers: (1) \code{eri_mean}; (2) \code{eri_min}; (3) \code{eri_max}; (4) \code{eri_az}.
}
\description{
Runs the Escape Route Index (ERI) algorithm for every 30 x 30m cell within an area of interest. ERI is the ratio between the distance one can travel within a given time frame, factoring in the surrounding landscape impediments, and the distance one could travel in that same time frame absent any impediments. It ranges from 0 (cannot evacuate from a particular location) to 1 (can evacuate with ideal speed).
}
\details{
\itemize{
\item In two dimensions, ERI is simple: the ratio between the distance one can travel through an existing landscape and the distance one could travel through an imaginary landscape without any impedance. If the former was 100 m and the latter was 200 m, ERI would be equal to 0.5.
\item However, ERI is calculated in two dimensions, as follows:
\enumerate{
\item Given a starting point...
\item An accumulative cost surface (representing travel time from that starting point) is generated using slope, elevation, and vegetation type-based transition matrix from \code{\link[=gen_tm]{gen_tm()}}.
\item A threshold is applied to that cost surface, defined by the evacuation time frame of interest, parameterized by \code{time}.
\item The edge of the resulting area represents a travel time isochrone (a line of equal travel time) from the starting point.
}
\item Accordingly, ERI can, and is, calculated in a number of ways:
\itemize{
\item ERImean is the mean distance to all points along the isochrone divided by the unimpeded distance one can travel in the same time frame, representing average evacuation ability in all directions
\item ERImin is the distance to the closest point along the isochrone divided by the unimpeded distance one can travel in the same time frame, representing evacuation ability in the worst-case scenario
\item ERImax is the distance to the furthest point along the isochrone divided by the unimpeded distance one can travel in the same time frame, representing evacuation ability in the best-case scenario
\item ERIaz is the azimuth between the starting point and the furthest point along the isochrone, representing the best evacuation direction
}
\item Note that, in order to avoid erroneous calculations at the edges of your AOI, \code{time} should be equal to or less than the same argument provided to both \code{\link[=download_lf]{download_lf()}} and \code{\link[=download_trans]{download_trans()}}.
\item Successful execution of this function will result in the saving of a file named "eri.tif" located within the user-defined \code{out_dir}.
}
}
\examples{
# read in aoi
aoi <- vect("C:/temp/study_area.shp")

# download landfire data using a 15 minute evacuation time
lf <- download_lf(aoi, 15, "C:/temp")

# access outputs
slope <- lf$slope
elev <- lf$elev
veg_type <- lf$veg_type

# download transportation data using a 15 minute evacuation time
trans <- download_trans(aoi, 15, "C:/temp")

# reclassify vegetation type
veg_type_rc <- reclass_veg(veg_type, trans, "C:/temp")

# generate transition matrix
tm <- gen_tm(slope, elev, veg_type_rc)

# calculate eri within the aoi using a 15 minute evacuation time
eri_area_result <- eri_area(veg_type_rc, tm, aoi, 15, 8, "C:/temp")
}
