% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eri_pt.R
\name{eri_pt}
\alias{eri_pt}
\title{Calculate ERI for a Single Point}
\usage{
eri_pt(veg_type, tm, start_pt, time, spatial = F)
}
\arguments{
\item{veg_type}{SpatRaster. The OpenStreetMap transportation data-augmented LANDFIRE vegetation type dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}, \code{\link[=download_trans]{download_trans()}}, and \code{\link[=reclass_veg]{reclass_veg()}}.}

\item{tm}{TransitionLayer. The transition matrix generated using \code{\link[=gen_tm]{gen_tm()}}.}

\item{start_pt}{SpatVector. The starting point from which evacuation will be simulated and ERI will be calculated.}

\item{time}{numeric. The total simulated evacuation time, in minutes, within which you aim to evaluate ERI.}

\item{spatial}{Boolean. Defines whether you only want quantitative ERI metrics (\code{spatial = FALSE}) or if you also want SpatVector/SpatRaster representations of the spatial features used to calculate ERI (\code{spatial = T}).}
}
\value{
If \code{spatial == FALSE}, a list with four numeric objects:
\itemize{
\item \code{eri_mean}
\item \code{eri_min}
\item \code{eri_max}
\item \code{eri_az}
}

If \code{spatial == TRUE}, all of the above, plus:
\itemize{
\item \code{start_pt}: SpatVector. The same starting point supplied by the user in the function call
\item \code{max_poly}: SpatVector. A polygon representing the distance one can travel in the absence of travel impediments. Takes the form of an octagon due to the 8-directional nature of cell adjacency in the accumulative cost simulations.
\item \code{act_poly}: SpatVector. A polygon representing the distance one can travel in the presence of travel impediments.
\item \code{tt_rast}: SpatRaster. The accumulative cost raster, clipped to the extent of \code{act_poly}, where each cell represents travel time in seconds from the starting point.
\item \code{pt_max}: SpatVector. A polygon representation of the travel time raster cell furthest from the starting point used as the basis of ERImax and ERIaz calculations.
\item \code{pt_min}: SpatVector. A polygon representation of the travel time raster cell closest from the starting point used as the basis of ERImin calculation.
}
}
\description{
Runs the Escape Route Index (ERI) algorithm for a single starting point. ERI is the ratio between the distance one can travel within a given time frame, factoring in the surrounding landscape impediments, and the distance one could travel in that same time frame absent any impediments. It ranges from 0 (cannot evacuate from a particular location) to 1 (can evacuate with ideal speed).
}
\details{
\itemize{
\item In two dimensions, ERI is simple: the ratio between the distance one can travel through an existing landscape and the distance one could travel through an imaginary landscape without any impedance. If the former was 100 m and the latter was 200 m, ERI would be equal to 0.5.
\item However, ERI is calculated in two dimensions, as follows:
\enumerate{
\item Given a starting point...
\item An accumulative cost surface (representing travel time from that starting point) is generated using slope, elevation, and vegetation type-based transition matrix from \code{\link[=gen_tm]{gen_tm()}}.
\item A threshold is applied to that cost surface, defined by the evacuation time frame of interest, parameterized by \code{time}.
\item The edge of the resulting area represents a travel time isochrone (a line of equal travel time) from the starting point.
}
\item Accordingly, ERI can, and is, calculated in a number of ways:
\itemize{
\item ERImean is the mean distance to all points along the isochrone divided by the unimpeded distance one can travel in the same time frame, representing average evacuation ability in all directions
\item ERImin is the distance to the closest point along the isochrone divided by the unimpeded distance one can travel in the same time frame, representing evacuation ability in the worst-case scenario
\item ERImax is the distance to the furthest point along the isochrone divided by the unimpeded distance one can travel in the same time frame, representing evacuation ability in the best-case scenario
\item ERIaz is the azimuth between the starting point and the furthest point along the isochrone, representing the best evacuation direction
}
\item The primary purpose of this function is to drive the \code{\link[=eri_area]{eri_area()}} function, which maps ERImean, ERImin, ERImax, and ERIaz across your entire AOI.
\item However, you can also run this on a single SpatVector starting point. If you intend to do so for a point of interest, you may consider setting \code{spatial = TRUE}. In addition to returning a list of numeric ERI metrics, this will return a series of SpatVector and SpatRaster objects that are useful for visualization and can be fed into \code{\link[=plot_eri]{plot_eri()}} for a simple visualization.
\item Note that, in order to avoid erroneous calculations at the edges of your AOI, \code{time} should be equal to or less than the same argument provided to both \code{\link[=download_lf]{download_lf()}} and \code{\link[=download_trans]{download_trans()}}.
}
}
\examples{
# read in aoi
aoi <- vect("C:/temp/study_area.shp")

# download landfire data using a 15 minute evacuation time
lf <- download_lf(aoi, 15, "C:/temp")

# access outputs
slope <- lf$slope
elev <- lf$elev
veg_type <- lf$veg_type

# download transportation data using a 15 minute evacuation time
trans <- download_trans(aoi, 15, "C:/temp")

# reclassify vegetation type
veg_type_rc <- reclass_veg(veg_type, trans, "C:/temp")

# generate transition matrix
tm <- gen_tm(slope, elev, veg_type_rc)

# generate a random starting point within the aoi
start_pt <- terra::project(aoi, veg_type_rc) |>
  terra::spatSample(1)

# calculate eri from that point using a 15 minute evacuation time
eri_pt_result <- eri_pt(veg_type_rc, tm, start_pt, 15, T)
}
