% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tm.R
\name{gen_tm}
\alias{gen_tm}
\title{Generate Transition Matrix}
\usage{
gen_tm(slope, elev, veg_type)
}
\arguments{
\item{slope}{SpatRaster. The LANDFIRE slope dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}.}

\item{elev}{SpatRaster. The LANDFIRE elevation dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}.}

\item{veg_type}{SpatRaster. The OpenStreetMap transportation data-augmented LANDFIRE vegetation type dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}, \code{\link[=download_trans]{download_trans()}}, and \code{\link[=reclass_veg]{reclass_veg()}}.}
}
\value{
A TransitionLayer representing the pedestrian conductance values (in meters per second) associated with every 8-directional adjacent cell pairing throughout the study area.
}
\description{
Generate a transition matrix using slope, elevation, and vegetation type data using the \code{\link[gdistance:transition]{gdistance::transition()}} function.
}
\details{
\itemize{
\item Travel rates, in meters, are calculated first through the use of the \code{\link[=slope_rate]{slope_rate()}} function, which estimates travel rate as a function of the terrain slope in the direction of movement.
\item Then, all slopes greater than 45 degrees, irrespective of travel direction, are treated as barriers.
\item The resulting travel rates are then multiplied by the vegetation type-based relative conductance factors, defined by \code{\link[=reclass_veg]{reclass_veg()}}.
}
}
\examples{
# read in aoi
aoi <- vect("C:/temp/study_area.shp")

# download landfire data using a 15 minute evacuation time
lf <- download_lf(aoi, 15, "C:/temp")

# access outputs
slope <- lf$slope
elev <- lf$elev
veg_type <- lf$veg_type

# download transportation data using a 15 minute evacuation time
trans <- download_trans(aoi, 15, "C:/temp")

# reclassify vegetation type
veg_type_rc <- reclass_veg(veg_type, trans, "C:/temp")

# generate transition matrix
tm <- gen_tm(slope, elev, veg_type_rc)
}
