% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eri_pt.R
\name{plot_eri_pt}
\alias{plot_eri_pt}
\title{Plot ERI for a Single Point}
\usage{
plot_eri_pt(elev, veg_type, eri_pt_result)
}
\arguments{
\item{elev}{SpatRaster. The LANDFIRE elevation dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}.}

\item{veg_type}{SpatRaster. The OpenStreetMap transportation data-augmented LANDFIRE vegetation type dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}, \code{\link[=download_trans]{download_trans()}}, and \code{\link[=reclass_veg]{reclass_veg()}}.}

\item{eri_pt_result}{list. The list returned from the successful execution of \code{\link[=eri_pt]{eri_pt()}}.}
}
\description{
Uses the outputs of \code{\link[=eri_pt]{eri_pt()}} to create a map that displays all of the spatial features that were used to calculate ERI.
}
\examples{
# read in aoi
aoi <- vect("C:/temp/study_area.shp")

# download landfire data using a 15 minute evacuation time
lf <- download_lf(aoi, 15, "C:/temp")

# access outputs
slope <- lf$slope
elev <- lf$elev
veg_type <- lf$veg_type

# download transportation data using a 15 minute evacuation time
trans <- download_trans(aoi, 15, "C:/temp")

# reclassify vegetation type
veg_type_rc <- reclass_veg(veg_type, trans, "C:/temp")

# generate transition matrix
tm <- gen_tm(slope, elev, veg_type_rc)

# generate a random starting point within the aoi
start_pt <- terra::project(aoi, veg_type_rc) |>
  terra::spatSample(1)

# calculate eri from that point using a 15 minute evacuation time
eri_pt_result <- eri_pt(veg_type_rc, tm, start_pt, 15, T)

# plot out the results
plot_eri_pt(elev, veg_type_rc, eri_pt_result)
}
