% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_time.R
\name{print_time}
\alias{print_time}
\title{Simple Date & Time Printer}
\usage{
print_time()
}
\value{
A character variable containing the current date and time.
}
\description{
Print the current date and time in a simple, clean format. Used in several functions for status printing.
}
\examples{
print_time()

}
