% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_veg.R
\name{reclass_veg}
\alias{reclass_veg}
\title{Reclassify Vegetation Data}
\usage{
reclass_veg(veg_type, trans, out_dir)
}
\arguments{
\item{veg_type}{SpatRaster. The LANDFIRE Existing Vegetation Type dataset derived from successful execution of \code{\link[=download_lf]{download_lf()}}.}

\item{trans}{SpatVector. The OpenStreetMap transportation dataset derived from successful execution of \code{\link[=download_trans]{download_trans()}}.}

\item{out_dir}{character. The output directory to which the resulting reclassified raster will be saved.}
}
\value{
A SpatRaster with cell values representing relative pedestrian conductance (see Details).
}
\description{
Combines the LANDFIRE Existing Vegetation Type and OpenStreetMap transportation data to create a single raster dataset representing the land cover-based relative conductance for pedestrian travel.
}
\details{
\itemize{
\item The conductance values modeled in this function were gleaned from a generalization of the pedestrian impedance factors used in the \href{https://wfdss.usgs.gov/wfdss_help/WFDSSHelp_Est_Grd_Medevac_Time.html}{WFDSS Estimated Ground Evacuation Time} layer. They are as follows:
\itemize{
\item Grass-dominated/non-burnable/transportation: 1x
\item Shrub-dominated: 0.5x
\item Tree-dominated: 0.25x
\item Water: 0x (acts as a barrier)
}
\item Note that this function will save the following TIF to \code{out_dir}: "veg_type_rc.tif".
}
}
\examples{
# read in aoi
aoi <- vect("C:/temp/study_area.shp")

# download landfire data using a 15 minute evacuation time
lf <- download_lf(aoi, 15, "C:/temp")

# access output
veg_type <- lf$veg_type

# download transportation data using a 15 minute evacuation time
trans <- download_trans(aoi, 15, "C:/temp")

# reclassify vegetation type
veg_type_rc <- reclass_veg(veg_type, trans, "C:/temp")
}
